Clean tests and minor refactoring WIP
