+++ http/netty/components/org.wso2.carbon.transport.http.netty/src/test/java/org/wso2/carbon/transport/http/netty/passthrough/PassThroughHttpGetMethodTestCase.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.transport.http.netty.passthrough.test;
+package org.wso2.carbon.transport.http.netty.passthrough;
 
 import io.netty.handler.codec.http.HttpMethod;
 import org.slf4j.Logger;
@@ -24,8 +24,11 @@ import org.slf4j.LoggerFactory;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
+import org.wso2.carbon.transport.http.netty.common.Constants;
 import org.wso2.carbon.transport.http.netty.config.ListenerConfiguration;
 import org.wso2.carbon.transport.http.netty.config.SenderConfiguration;
+import org.wso2.carbon.transport.http.netty.config.TransportsConfiguration;
+import org.wso2.carbon.transport.http.netty.config.YAMLTransportConfigurationBuilder;
 import org.wso2.carbon.transport.http.netty.listener.HTTPTransportListener;
 import org.wso2.carbon.transport.http.netty.util.TestUtil;
 import org.wso2.carbon.transport.http.netty.util.server.HTTPServer;
@@ -35,16 +38,17 @@ import java.net.HttpURLConnection;
 import java.net.URI;
 
 import static org.testng.AssertJUnit.assertEquals;
-import static org.testng.AssertJUnit.assertTrue;
 
 /**
  * A test class for passthrough transport
  */
-public class PassThroughHttpPOSTMethodTestCase {
+public class PassThroughHttpGetMethodTestCase {
 
-    protected static final Logger LOGGER = LoggerFactory.getLogger(PassThroughHttpPOSTMethodTestCase.class);
+    private static final Logger log = LoggerFactory.getLogger(PassThroughHttpGetMethodTestCase.class);
 
     private HTTPTransportListener httpTransportListener;
+
+    private static final String testValue = "Test Message";
 
     private HTTPServer httpServer;
 
@@ -54,42 +58,31 @@ public class PassThroughHttpPOSTMethodTestCase {
 
     private URI baseURI = URI.create(String.format("http://%s:%d", "localhost", 8490));
 
-    @BeforeClass(groups = "passthroughPost",
-                 dependsOnGroups = "passthroughGET")
+    @BeforeClass
     public void setUp() {
-        listenerConfiguration = new ListenerConfiguration();
-        listenerConfiguration.setHost(TestUtil.TEST_HOST);
-        listenerConfiguration.setId("test-listener");
-        listenerConfiguration.setPort(TestUtil.TEST_ESB_PORT);
-        senderConfiguration = new SenderConfiguration("passthrough-sender");
-        httpTransportListener = TestUtil
-                .startCarbonTransport(listenerConfiguration, senderConfiguration, new PassthroughMessageProcessor());
-        httpServer = TestUtil.startHTTPServer(TestUtil.TEST_SERVER_PORT);
+        TransportsConfiguration configuration = YAMLTransportConfigurationBuilder
+                .build("src/test/resources/simple-test-config/netty-transports.yml");
+        httpTransportListener = TestUtil.startCarbonTransport(configuration, new PassthroughMessageProcessor());
+        httpServer = TestUtil.startHTTPServer(TestUtil.TEST_SERVER_PORT, testValue, Constants.TEXT_PLAIN);
     }
 
-    public PassThroughHttpPOSTMethodTestCase() {
-        super();
-    }
-
-    @Test(groups = "passthroughPost",
-          dependsOnGroups = "passthroughGET")
-    public void passthroughPOSTTestCase() {
-        String testValue = "Test Message";
+    @Test
+    public void passthroughGetTest() {
         try {
-            HttpURLConnection urlConn = TestUtil.request(baseURI, "/", HttpMethod.POST.name(), true);
-            TestUtil.writeContent(urlConn, testValue);
+            HttpURLConnection urlConn = TestUtil.request(baseURI, "/", HttpMethod.GET.name(), true);
             String content = TestUtil.getContent(urlConn);
             assertEquals(testValue, content);
             urlConn.disconnect();
         } catch (IOException e) {
-            LOGGER.error("IO Exception occurred", e);
-            assertTrue(false);
+            log.error("IO Exception occurred", e);
         }
-
     }
-    
-    @AfterClass(groups = "passthroughPost",
-                dependsOnGroups = "passthroughGET")
+
+    public PassThroughHttpGetMethodTestCase() {
+        super();
+    }
+
+    @AfterClass(groups = "passthroughGET")
     public void cleanUp() {
         TestUtil.cleanUp(httpTransportListener, httpServer);
     }
