+++ http/netty/components/org.wso2.carbon.transport.http.netty/src/test/resources/simple-test-config/netty-transports.yml
@@ -1,0 +1,150 @@
+################################################################################
+#   Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+################################################################################
+transportProperties:
+ -
+  name: "latency.metrics.enabled"
+  value: true
+ -
+  name: "request.size.validation"
+  value: false
+ -
+  name: "request.size.validation.maximum.value" #in bytes
+  value: 2048
+ -
+  name: "request.size.validation.reject.status.code"
+  value: 401
+ -
+  name: "request.size.validation.reject.message"
+  value: "Message is too big, hence cannot proceed"
+ -
+  name: "request.size.validation.reject.message.content.type"
+  value: "plain/text"
+ -
+  name: "header.size.validation"
+  value: false
+ -
+  name: "header.validation.maximum.request.line" #in bytes
+  value: 2048
+ -
+  name: "header.validation.maximum.size" #in bytes
+  value: 2048
+ -
+  name: "header.validation.maximum.chunk.size" #in bytes, refere http://netty.io/4.0/api/io/netty/handler/codec/http/HttpRequestDecoder.html
+  value: 2048
+ -
+  name: "header.validation.reject.status.code"
+  value: 401
+ -
+  name: "header.validation.reject.message"
+  value: "Message's header size is too big, hence cannot proceed"
+ -
+  name: "header.validation.reject.message.content.type"
+  value: "plain/text"
+ -
+  name: "server.bootstrap.boss.group.size"
+  value: 4
+ -
+  name: "server.bootstrap.worker.group.size"
+  value: 8
+ -
+  name: "server.bootstrap.nodelay"
+  value: true
+ -
+  name: "server.bootstrap.keepalive"
+  value: true
+ -
+  name: "server.bootstrap.sendbuffersize"
+  value: 1048576
+ -
+  name: "server.bootstrap.recievebuffersize"
+  value: 1048576
+ -
+  name: "server.bootstrap.connect.timeout"
+  value: 15000
+ -
+  name: "server.bootstrap.socket.reuse"
+  value: false
+ -
+  name: "server.bootstrap.socket.timeout"
+  value: 15
+ -
+  name: "server.bootstrap.socket.backlog"
+  value: 100
+
+listenerConfigurations:
+ -
+  id: "default"
+  host: "0.0.0.0"
+  port: 8490
+
+senderConfigurations:
+ -
+  id: "http-sender"
+  parameters:
+    -
+     name: "connection.pool.count"
+     value: 0
+    -
+     name: "client.bootstrap.nodelay"
+     value: true
+    -
+     name: "client.bootstrap.keepalive"
+     value: true
+    -
+     name: "client.bootstrap.sendbuffersize"
+     value: 1048576
+    -
+     name: "client.bootstrap.recievebuffersize"
+     value: 1048576
+    -
+     name: "client.bootstrap.connect.timeout"
+     value: 15000
+    -
+     name: "client.bootstrap.socket.reuse"
+     value: false
+    -
+     name: "client.bootstrap.socket.timeout"
+     value: 15
+ -
+  id: "https-sender"
+  scheme: "https"
+  trustStoreFile: "simple-test-config/client-truststore.jks"
+  trustStorePass: "wso2carbon"
+  parameters:
+    -
+     name: "connection.pool.count"
+     value: 0
+    -
+     name: "client.bootstrap.nodelay"
+     value: true
+    -
+     name: "client.bootstrap.keepalive"
+     value: true
+    -
+     name: "client.bootstrap.sendbuffersize"
+     value: 1048576
+    -
+     name: "client.bootstrap.recievebuffersize"
+     value: 1048576
+    -
+     name: "client.bootstrap.connect.timeout"
+     value: 15000
+    -
+     name: "client.bootstrap.socket.reuse"
+     value: false
+    -
+     name: "client.bootstrap.socket.timeout"
+     value: 15
